+++ src/pages/settings/preload.js
@@ -10,7 +10,8 @@ let trayIcon,
   port,
   menuBar,
   mutedArtists,
-  singleInstance;
+  singleInstance,
+  disableHardwareMediaKeys;
 
 const { store, settings } = require("./../../scripts/settings");
 const { ipcRenderer } = require("electron");
@@ -34,6 +35,7 @@ function refreshSettings() {
   muteArtists.checked = store.get(settings.muteArtists);
   mutedArtists.value = store.get(settings.mutedArtists).join("\n");
   singleInstance.checked = store.get(settings.singleInstance);
+  disableHardwareMediaKeys.checked = store.get(settings.disableHardwareMediaKeys);
 }
 
 /**
@@ -114,6 +116,7 @@ window.addEventListener("DOMContentLoaded", () => {
   muteArtists = get("muteArtists");
   mutedArtists = get("mutedArtists");
   singleInstance = get("singleInstance");
+  disableHardwareMediaKeys = get("disableHardwareMediaKeys");
 
   refreshSettings();
 
@@ -130,4 +133,5 @@ window.addEventListener("DOMContentLoaded", () => {
   addInputListener(muteArtists, settings.muteArtists);
   addTextAreaListener(mutedArtists, settings.mutedArtists);
   addInputListener(singleInstance, settings.singleInstance);
+  addInputListener(disableHardwareMediaKeys, settings.disableHardwareMediaKeys);
 });